Fix services (delete operation)

+++ src/main/java/com/upp/apteka/service/impl/DoctorServiceImpl.java
@@ -46,6 +46,10 @@ public class DoctorServiceImpl implements DoctorService{
 
 	//@Override
 	public boolean delete(Long key) {
+		Doctor doctor = read(key);
+		
+		if(doctor.getPrescriptions().size() != 0)
+			return false;
 		return doctorRepository.delete(key);
 	}
